+++ src/cartesian-chart/interfaces.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import * as CoreTypes from "../core/interfaces";
-import { NonCancelableEventHandler } from "../internal/events";
+import { NonCancelableEventHandler } from "../types/events";
 
 // CartesianChartProps is the type for CartesianChart React component properties. Unlike in Highcharts API,
 // we pass options directly to the component, instead of grouping them all into a single "options" property.
